pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["pavlenko <mozguru@mail.ru>"]
readme = "README.md"
packages = [{ include = "gendiff" },]


[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.dev-dependencies]
flake8 = "^7.0.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pyyaml = "6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

